migrate redundancy to multisupervisor
